services/training/requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
joblib
matplotlib
seaborn
pyyaml
